+++ package.json
@@ -14,16 +14,17 @@
   "devDependencies": {
     "@babel/core": "7.12.13",
     "@babel/preset-typescript": "7.12.13",
-    "@nrwl/eslint-plugin-nx": "^15.8.7",
-    "@nrwl/jest": "^15.8.7",
-    "@nrwl/js": "^15.8.7",
-    "@nrwl/linter": "^15.8.7",
-    "@nrwl/react": "^15.8.7",
-    "@nrwl/rollup": "^15.8.7",
-    "@nrwl/storybook": "^15.8.7",
-    "@nrwl/vite": "^15.8.7",
-    "@nrwl/webpack": "^15.8.7",
-    "@nrwl/workspace": "^15.8.7",
+    "@nrwl/esbuild": "^15.9.2",
+    "@nrwl/eslint-plugin-nx": "^15.9.2",
+    "@nrwl/jest": "^15.9.2",
+    "@nrwl/js": "^15.9.2",
+    "@nrwl/linter": "^15.9.2",
+    "@nrwl/react": "^15.9.2",
+    "@nrwl/rollup": "^15.9.2",
+    "@nrwl/storybook": "^15.9.2",
+    "@nrwl/vite": "^15.9.2",
+    "@nrwl/webpack": "^15.9.2",
+    "@nrwl/workspace": "^15.9.2",
     "@storybook/addon-essentials": "^6.5.15",
     "@storybook/builder-vite": "^0.2.6",
     "@storybook/builder-webpack5": "^6.5.15",
@@ -49,7 +50,7 @@
     "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.4.1",
     "jest-file-snapshot": "^0.5.0",
-    "nx": "^15.8.7",
+    "nx": "^15.9.2",
     "prettier": "^2.6.2",
     "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
