+++ packages/data/project.json
@@ -12,7 +12,7 @@
         "project": "packages/data/package.json",
         "main": "packages/data/src/index.ts",
         "tsConfig": "packages/data/tsconfig.lib.json",
-        "assets": ["packages/data/src/definitions.js"],
+        "assets": [],
         "format": ["esm", "cjs"]
       }
     },
