+++ packages/data/src/actions.ts
@@ -1,6 +1,6 @@
 import React from "react";
 import { useBatiq } from "@batiq/expo-runtime";
-import { importDataSourceModule } from "@batiq/shared";
+import { importDataSourceModule } from "@batiq/import-helper";
 
 export const authenticate = () => {
   const batiq = useBatiq();
@@ -11,6 +11,20 @@ export const authenticate = () => {
       const data = await importDataSourceModule(schema, datasourceName);
 
       return data.authenticate(batiq, params);
+    },
+    [schema]
+  );
+};
+
+export const logout = () => {
+  const batiq = useBatiq();
+  const schema = batiq.getSchema();
+
+  return React.useCallback(
+    async (datasourceName: string) => {
+      const data = await importDataSourceModule(schema, datasourceName);
+
+      return data.logout(batiq);
     },
     [schema]
   );
