+++ packages/data/src/index.ts
@@ -1,1 +1,2 @@
 export { OpenAPI } from "./OpenAPI";
+export { authenticate, logout } from "./actions";
