+++ packages/expo-runtime/project.json
@@ -14,6 +14,12 @@
         "tsConfig": "packages/expo-runtime/tsconfig.lib.json",
         "assets": [],
         "format": ["esm", "cjs"],
+        "external": [
+          "@react-navigation/native",
+          "expo",
+          "@react-navigation/bottom-tabs",
+          "@react-navigation/native-stack"
+        ],
         "rollupConfig": "packages/expo-runtime/rollup.config.js"
       }
     },
