+++ packages/expo-runtime/src/Query.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import useSwr from "swr/immutable";
 import { useBatiq } from "./AppContext";
-import { importDataSourceModule } from "@batiq/shared";
+import { importDataSourceModule } from "@batiq/import-helper";
 import { withData } from "./DataProvider";
 
 type Props = {
