+++ packages/runtime/src/PageRuntime.tsx
@@ -5,7 +5,7 @@ import {
   Component as ComponentIR,
   ComponentImport,
 } from "@batiq/ir";
-import { importModule } from "@batiq/shared";
+import { importModule } from "@batiq/import-helper";
 import { valueToRuntime } from "./utils/valueToRuntime";
 import { withComponentProvider } from "@batiq/expo-runtime";
