--- packages/shared/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "@batiq/shared",
-  "version": "0.1.0",
-  "type": "module",
-  "devDependencies": {}
-}
