--- packages/shared/project.json
@@ -1,42 +1,0 @@
-{
-  "name": "shared",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/shared/src",
-  "projectType": "library",
-  "implicitDependencies": [
-    "!actions",
-    "!components",
-    "!core",
-    "!data",
-    "!expression",
-    "!expo-runtime"
-  ],
-  "targets": {
-    "build": {
-      "executor": "@nrwl/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/packages/shared",
-        "main": "packages/shared/src/index.ts",
-        "tsConfig": "packages/shared/tsconfig.lib.json",
-        "buildableProjectDepsInPackageJsonType": "peerDependencies",
-        "assets": []
-      }
-    },
-    "publish": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "node tools/scripts/publish.mjs shared{args.ver} {args.tag}"
-      },
-      "dependsOn": ["build"]
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["packages/shared/**/*.ts"]
-      }
-    }
-  },
-  "tags": []
-}
