--- packages/shared/src/index.ts
@@ -1,2 +1,0 @@
-export { importModule, importNamedModule } from "./importModule";
-export { importDataSourceModule } from "./importDataSourceModule";
