+++ packages/utils/import-helper/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@batiq/import-helper",
+  "version": "0.1.0",
+  "type": "commonjs"
+}
