+++ packages/utils/import-helper/src/index.ts
@@ -1,0 +1,6 @@
+export {
+  setImportModule,
+  importModule,
+  importNamedModule,
+} from "./importModule";
+export { importDataSourceModule } from "./importDataSourceModule";
